+++ CHANGELOG.md
@@ -1,4 +1,12 @@
+# v0.0.5
+
+-   Added new plugins
+    - `NSFW Detection` - Detect NSFW content in images.
+    -   `ObjectDetection` - Detect bounding boxes of objects in the image.
+    -   `ViewDetection` - Classifies wear type and view type of products in the image
+
+
 # v0.0.4
 
--   Transformations are updated to their latest API
--   Added support for GCS signed URL uploads
+- Transformations are updated to their latest API
+- Added support for GCS signed URL uploads